created all the REST endpoints for /companies

+++ src/main/java/com/paraett/companiesservice/service/CompanyService.java
@@ -1,9 +1,13 @@
 package com.paraett.companiesservice.service;
 
+import com.paraett.companiesservice.exception.CompanyNotFoundException;
 import com.paraett.companiesservice.model.dtos.CompanyRegisterDto;
 import com.paraett.companiesservice.model.entities.Company;
 import com.paraett.companiesservice.repository.CompanyRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CompanyService {
@@ -21,6 +25,34 @@ public class CompanyService {
     }
 
     public void deleteCompany(Long id) {
-        companyRepository.deleteById(id);
+        Optional<Company> optCompany = this.companyRepository.findById(id);
+
+        if (optCompany.isPresent()) {
+            companyRepository.deleteById(id);
+            return;
+        }
+        throw new CompanyNotFoundException("id-" + id);
+    }
+
+    public Company updateCompany(Long id, Company company) {
+        Optional<Company> optCompany = this.companyRepository.findById(id);
+
+        if (optCompany.isPresent()) {
+            company.setId(optCompany.get().getId());
+            return companyRepository.save(company);
+        }
+        throw new CompanyNotFoundException("id-" + id);
+    }
+
+    public Company getCompany(Long id) {
+        Optional<Company> optCompany = this.companyRepository.findById(id);
+        if (optCompany.isPresent()) {
+            return optCompany.get();
+        }
+        throw new CompanyNotFoundException("id-" + id);
+    }
+
+    public List<Company> getAllCompanies() {
+        return this.companyRepository.findAll();
     }
 }
